refactor: move 'admin' module into 'domain' directory
