+++ src/app.module.ts
@@ -3,7 +3,7 @@ import { PrismaModule } from "./infra/prisma/prisma.module";
 import { ConfigModule } from "@nestjs/config";
 import { UserModule } from "./domain/user/user.module";
 import { AuthModule } from "./infra/auth/auth.module";
-import { AdminModule } from "./admin/admin.module";
+import { AdminModule } from "./domain/admin/admin.module";
 import { HealthModule } from "./health/health.module";
 
 @Module({
