+++ src/domain/admin/controllers/v1/dto/swagger/admin-find-all-api-response.dto.ts
